restructured frontend folders
