--- frontend/src/WaitingRoom.jsx
@@ -1,36 +1,0 @@
-import React, { useState } from "react";
-import { Link } from "react-router-dom";
-import "./WaitingRoom.css";
-import { baseURL, gameID } from "./Welcome";
-
-export default function Room({ initialhand, setinitialhand, setisgamestarted }) {
-
-    const startGame = () => {
-      console.log(gameID);
-      let url = baseURL + "startgame/" + gameID + "/"
-      console.log(url);
-      fetch(url, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json"
-      }})
-        .then(response => response.json())
-        .then(data => {
-          setinitialhand(data.Players.John.hand); 
-          console.log(initialhand);
-          setisgamestarted(true); // Set game started state to true
-        })
-        .catch(error => {
-          alert(error);
-          console.error("Error: ", error);
-        })
-    }
-
-    return (
-        <div className="room">
-            <button onClick={startGame}>
-              Start Game
-            </button>
-        </div>
-    )
-}
