+++ frontend/src/css/Infoboard.css
