+++ frontend/src/css/WaitingRoom.css
