+++ frontend/src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter, Route, Routes, Navigate } from "react-router-dom";
-import Welcome from './Welcome.jsx'
+import Welcome from './pages/Welcome.jsx'
 import App from './App.jsx'
 import './index.css'
